dtype update

+++ sklearnex/spmd/basic_statistics/tests/test_incremental_basic_statistics_spmd.py
@@ -60,9 +60,9 @@ def test_incremental_basic_statistics_fit_spmd_gold(dataframe, queue, weighted, 
             [0.0, 5.0, 32.0],
             [0.0, 6.0, 64.0],
             [0.0, 7.0, 128.0],
-        ]
-    ).astype(dtype=dtype)
-
+        ],
+        dtype=dtype,
+    )
     dpt_data = _convert_to_dataframe(data, sycl_queue=queue, target_df=dataframe)
 
     local_dpt_data = _convert_to_dataframe(
@@ -70,7 +70,7 @@ def test_incremental_basic_statistics_fit_spmd_gold(dataframe, queue, weighted, 
     )
 
     if weighted:
-        weights = np.array([0, 1, 2, 3, 4, 5, 6, 7]).astype(dtype=dtype)
+        weights = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0], dtype=dtype)
         dpt_weights = _convert_to_dataframe(
             weights, sycl_queue=queue, target_df=dataframe
         )
@@ -123,15 +123,15 @@ def test_incremental_basic_statistics_partial_fit_spmd_gold(
             [0.0, 5.0, 32.0],
             [0.0, 6.0, 64.0],
             [0.0, 7.0, 128.0],
-        ]
+        ],
+        dtype=dtype,
     )
-    data = data.astype(dtype=dtype)
     dpt_data = _convert_to_dataframe(data, sycl_queue=queue, target_df=dataframe)
     local_data = _get_local_tensor(data)
     split_local_data = np.array_split(local_data, num_blocks)
 
     if weighted:
-        weights = np.array([0, 1, 2, 3, 4, 5, 6, 7])
+        weights = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0], dtype=dtype)
         dpt_weights = _convert_to_dataframe(
             weights, sycl_queue=queue, target_df=dataframe
         )
@@ -192,15 +192,15 @@ def test_incremental_basic_statistics_single_option_partial_fit_spmd_gold(
             [0.0, 5.0, 32.0],
             [0.0, 6.0, 64.0],
             [0.0, 7.0, 128.0],
-        ]
+        ],
+        dtype=dtype,
     )
-    data = data.astype(dtype=dtype)
     dpt_data = _convert_to_dataframe(data, sycl_queue=queue, target_df=dataframe)
     local_data = _get_local_tensor(data)
     split_local_data = np.array_split(local_data, num_blocks)
 
     if weighted:
-        weights = np.array([0, 1, 2, 3, 4, 5, 6, 7])
+        weights = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0], dtype=dtype)
         dpt_weights = _convert_to_dataframe(
             weights, sycl_queue=queue, target_df=dataframe
         )
